Experiment to make initial page load faster...
Some pro tips from vercel n co
