+++ apps/docs/app/(diffshub)/(view)/_components/utils.ts
@@ -93,7 +93,7 @@ export function getPatchViewerHref(input: string): string | undefined {
     const githubPath = getGitHubPathFromURL(parsedURL);
     if (githubPath != null) return githubPath;
     if (parsedURL.pathname !== '/') {
-      return `${parsedURL.pathname}?domain=${encodeURIComponent(parsedURL.hostname)}`;
+      return getAlternateDomainViewerHref(parsedURL);
     }
     return undefined;
   } catch {
@@ -111,7 +111,7 @@ export function getPatchViewerHref(input: string): string | undefined {
       const githubPath = getGitHubPathFromURL(parsedURL);
       if (githubPath != null) return githubPath;
       if (parsedURL.pathname !== '/') {
-        return `${parsedURL.pathname}?domain=${encodeURIComponent(parsedURL.hostname)}`;
+        return getAlternateDomainViewerHref(parsedURL);
       }
     } catch {
       // Not parseable even with https:// prefix.
@@ -173,6 +173,10 @@ export function resolveDiffshubViewerRoute(
     upstreamPath,
     url: `https://${host}${upstreamPath}`,
   };
+}
+
+function getAlternateDomainViewerHref(parsedURL: URL): string {
+  return `${parsedURL.pathname}?domain=${encodeURIComponent(parsedURL.hostname)}`;
 }
 
 function getGitHubPathFromURL(parsedURL: URL): string | undefined {
