+++ apps/docs/lib/diffshubPatchTypes.ts
@@ -1,0 +1,7 @@
+export interface InitialDiffshubPatchResponse {
+  body: ReadableStream<string> | null;
+  bodyText: string | null;
+  ok: boolean;
+  status: number;
+  statusText: string;
+}
